# docs/openapi-identity.yaml
openapi: 3.1.0
info:
  title: Pulso Piura API
  version: 0.8.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /organizations:
    get:
      summary: Lista únicamente las organizaciones activas del usuario
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Organizaciones y rol contextual}
    post:
      summary: Crea una organización para un dueño aprobado o administrador global
      security: [{ bearerAuth: [] }]
      responses:
        "201": {description: Organización creada}
        "400": {description: Nombre inválido}
        "403": {description: La capacidad de dueño de cancha no está aprobada}
  /organizations/{organizationId}:
    parameters:
      - in: path
        name: organizationId
        required: true
        schema: {type: string, format: uuid}
    get:
      summary: Consulta una organización con membresía activa
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Organización accesible}
        "403": {description: Sin acceso al tenant solicitado}
        "404": {description: Organización no encontrada}
  /organizations/{organizationId}/invitations:
    post:
      summary: Invita un administrador u operador por correo
      security: [{ bearerAuth: [] }]
      responses:
        "201": {description: Invitación pendiente creada}
        "400": {description: Correo o rol inválido, o invitación duplicada}
        "403": {description: Sin permiso para administrar miembros}
  /organizations/invitations/{invitationId}/accept:
    post:
      summary: Acepta una invitación cuyo correo coincide con el usuario autenticado
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Membresía activada}
        "403": {description: La invitación pertenece a otro correo}
        "404": {description: Invitación no encontrada}
  /organizations/{organizationId}/members:
    get:
      summary: Lista miembros activos para propietarios y administradores
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Miembros activos}
        "403": {description: Sin permiso o acceso a otro tenant}
  /organizations/{organizationId}/members/{userId}:
    delete:
      summary: Revoca una membresía activa, excepto la del propietario
      security: [{ bearerAuth: [] }]
      responses:
        "204": {description: Acceso revocado}
        "400": {description: No se puede revocar al propietario}
        "403": {description: Sin permiso o acceso a otro tenant}
        "404": {description: Membresía activa no encontrada}
...
  /matches:
    get:
      summary: Lista partidos públicos publicados y futuros
      parameters:
        - {in: query, name: sport, required: false, schema: {type: string}}
      responses:
        "200": {description: Partidos disponibles, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/Match"}}}}}
    post:
      summary: Crea un partido desde una reserva propia confirmada
      security: [{ bearerAuth: [] }]
      responses:
        "201": {description: Partido creado en borrador, content: {application/json: {schema: {$ref: "#/components/schemas/Match"}}}}
        "401": {description: Token ausente o inválido}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /matches/mine:
    get:
      summary: Lista únicamente los partidos publicados por el organizador autenticado
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Partidos propios, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/Match"}}}}}
        "401": {description: Token ausente o inválido}
        "403": {$ref: "#/components/responses/Forbidden"}
...
  /matches/participations/me:
    get:
      summary: Lista las inscripciones activas del jugador para la sección Actividad
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Inscripciones del jugador, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/MatchActivity"}}}}}
        "401": {description: Token ausente o inválido}
...
  /platform/summary:
    get:
      summary: Devuelve indicadores globales para la consola de plataforma
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Resumen global}
        "403": {description: Requiere PLATFORM_ADMIN}
  /platform/users:
    get:
      summary: Lista las cuentas y capacidades registradas
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Usuarios de plataforma}
        "403": {description: Requiere PLATFORM_ADMIN}
  /platform/capability-requests:
    get:
      summary: Lista solicitudes enriquecidas con nombre y correo
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Solicitudes de capacidad}
        "403": {description: Requiere PLATFORM_ADMIN}
  /platform/capability-requests/{requestId}/review:
    parameters:
      - {in: path, name: requestId, required: true, schema: {type: string, format: uuid}}
    post:
      summary: Aprueba o rechaza una solicitud pendiente
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Solicitud revisada}
        "403": {description: Requiere PLATFORM_ADMIN}
  /platform/capability-requests/{requestId}/revoke:
    parameters:
      - {in: path, name: requestId, required: true, schema: {type: string, format: uuid}}
    post:
      summary: Revoca una capacidad aprobada
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Capacidad revocada}
        "403": {description: Requiere PLATFORM_ADMIN}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Forbidden: {description: Sin membresía o permiso en la organización, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
...
    Conflict: {description: Estado, versión o dato en conflicto, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
  schemas:
    Problem:
      type: object
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
...
    Match:
      type: object
      required: [id, publicSlug, sportSpaceId, spaceName, venueName, venueAddress, title, sportCode, formatCode, skillLevel, minPlayers, maxPlayers, occupiedPlayers, availablePlayers, priceMinor, currency, visibility, cancellationPolicy, startsAt, endsAt, status, participantPreview, version]
      properties:
        id: {type: string, format: uuid}
        publicSlug: {type: string}
        sportSpaceId: {type: string, format: uuid}
        spaceName: {type: string}
        venueName: {type: string}
        venueAddress: {type: string}
        title: {type: string}
        sportCode: {type: string}
        formatCode: {type: string}
        skillLevel: {type: string, enum: [ALL_LEVELS, BEGINNER, INTERMEDIATE, ADVANCED]}
        minPlayers: {type: integer, minimum: 1}
        maxPlayers: {type: integer, minimum: 1}
        occupiedPlayers: {type: integer, minimum: 0}
        availablePlayers: {type: integer, minimum: 0}
        priceMinor: {type: integer, format: int64, minimum: 0}
        currency: {type: string, const: PEN}
        visibility: {type: string, enum: [PUBLIC, LINK, PRIVATE]}
        cancellationPolicy: {type: string}
        startsAt: {type: string, format: date-time}
        endsAt: {type: string, format: date-time}
        status: {type: string, enum: [DRAFT, PUBLISHED, CANCELLED]}
        participantPreview:
          type: array
          maxItems: 7
          items: {$ref: "#/components/schemas/ParticipantPreview"}
        version: {type: integer, format: int64, minimum: 0}
    ParticipantPreview:
      type: object
      required: [displayName]
      properties:
        displayName: {type: string, maxLength: 120}
        avatarUrl: {type: [string, "null"], maxLength: 500}
...
    MatchActivity:
      type: object
      required: [id, publicSlug, title, sportCode, formatCode, spaceName, venueName, venueAddress, startsAt, endsAt, participationStatus, paymentStatus, paidMinor, currency]
      properties:
        id: {type: string, format: uuid}
        publicSlug: {type: string}
        title: {type: string}
        sportCode: {type: string}
        formatCode: {type: string}
        spaceName: {type: string}
        venueName: {type: string}
        venueAddress: {type: string}
        startsAt: {type: string, format: date-time}
        endsAt: {type: string, format: date-time}
        participationStatus: {type: string, enum: [JOINED, WAITLISTED]}
        paymentStatus: {type: string, enum: [PAID, PENDING, UNPAID, NOT_REQUIRED]}
        paidMinor: {type: integer, format: int64, minimum: 0}
        currency: {type: string, const: PEN}
...